% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{tda_clpyr}
\alias{tda_clpyr}
\title{Pyramidal clustering}
\usage{
tda_clpyr(d, linkage = c("single", "complete"), ...)
}
\arguments{
\item{d}{symmetric distance matrix.}

\item{linkage}{"single" or "complete" link.}

\item{...}{passed to \code{\link{tda_run}}.}
}
\value{
the printed output.
}
\description{
TDA's \code{clpyr} on a distance matrix.
}
\examples{
r <- tda_clpyr(as.matrix(dist(c(0, 1, 5, 6))))
# the top of the pyramid joins everything: 4 objects spanning 1..4
tail(r$pyramid, 1)
}
