% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-analysis.R
\name{tda_gpro}
\alias{tda_gpro}
\title{Projection of proximities}
\usage{
tda_gpro(d, max_iterations = 50, ...)
}
\arguments{
\item{d}{symmetric proximity matrix.}

\item{max_iterations}{iteration cap.}

\item{...}{passed to \code{\link{tda_run}}.}
}
\value{
\code{projection}, the projected coordinates, and
  \code{output}.
}
\description{
TDA's \code{gpro} on a proximity matrix given as an undirected
valued graph.
}
\examples{
r <- tda_gpro(as.matrix(dist(cbind(c(0, 0, 3, 3), c(0, 2, 0, 2)))))
r$projection
}
