% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{tda_mpinv}
\alias{tda_mpinv}
\title{Invert a permutation}
\usage{
tda_mpinv(p, dir = tempfile("tda"))
}
\arguments{
\item{p}{an integer permutation of \code{1:length(p)}.}

\item{dir}{working directory.}
}
\value{
An integer vector, the inverse permutation.
}
\description{
TDA's \code{mpinv}: given a permutation \code{p} of \code{1:n},
returns \code{q} such that \code{q[p[i]] == i} -- matching base
R's \code{order(p)}.
}
\examples{
tda_mpinv(c(3, 1, 4, 2))   # == order(c(3, 1, 4, 2))
}
\seealso{
Other matrix algebra:
\code{\link[=tda_mcel]{tda_mcel()}},
\code{\link[=tda_mcent]{tda_mcent()}},
\code{\link[=tda_mch]{tda_mch()}},
\code{\link[=tda_mcross]{tda_mcross()}},
\code{\link[=tda_mdiag]{tda_mdiag()}},
\code{\link[=tda_mev]{tda_mev()}},
\code{\link[=tda_mevs]{tda_mevs()}},
\code{\link[=tda_midf]{tda_midf()}},
\code{\link[=tda_midf1]{tda_midf1()}},
\code{\link[=tda_midf2]{tda_midf2()}},
\code{\link[=tda_midf3]{tda_midf3()}},
\code{\link[=tda_minvs]{tda_minvs()}},
\code{\link[=tda_mkmet]{tda_mkmet()}},
\code{\link[=tda_mkp]{tda_mkp()}},
\code{\link[=tda_mmul]{tda_mmul()}},
\code{\link[=tda_mnc]{tda_mnc()}},
\code{\link[=tda_mnrow]{tda_mnrow()}},
\code{\link[=tda_mpbl]{tda_mpbl()}},
\code{\link[=tda_mpfit]{tda_mpfit()}},
\code{\link[=tda_mpit]{tda_mpit()}},
\code{\link[=tda_mple]{tda_mple()}},
\code{\link[=tda_mpz]{tda_mpz()}},
\code{\link[=tda_mscal1]{tda_mscal1()}},
\code{\link[=tda_msqrtd]{tda_msqrtd()}},
\code{\link[=tda_msvd]{tda_msvd()}},
\code{\link[=tda_mwvec]{tda_mwvec()}}
}
\concept{matrix algebra}
