% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tda_ndvar}
\alias{tda_ndvar}
\title{Dummy variables}
\usage{
tda_ndvar(data, ...)
}
\arguments{
\item{data}{data frame of integer-valued variables.}

\item{...}{passed to \code{\link{tda_run}}.}
}
\value{
the printed output describing the created dummies.
}
\description{
TDA's \code{ndvar}: expands variables into dummies. Verified in
the suite against \code{model.matrix}.
}
\examples{
# one dummy per observed level of edu, named edu_1, edu_2, edu_3
tda_ndvar(data.frame(edu = c(1, 2, 3, 2, 1)))
}
