% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcmds.R
\name{tda_plglob3}
\alias{tda_plglob3}
\title{Globe in a 3-d plot}
\usage{
tda_plglob3(radius, lon = seq(-150, 150, 30), lat = seq(-60, 60, 30), ...)
}
\arguments{
\item{radius}{sphere radius.}

\item{lon, lat}{grid sequences (degrees).}

\item{...}{passed to \code{\link{tda_run}}.}
}
\value{
path of the PostScript file, invisibly.
}
\description{
TDA's \code{plglob3}: longitude/latitude grid on a sphere.
}
\examples{
# a globe with meridians every 30 degrees and parallels at -60..60
f <- tda_plglob3(1)
tda_plot_ps(tda_read_ps(f))

# a coarser grid, as one command of a session
p <- tda_ps3(xlim = c(-2, 2), ylim = c(-2, 2), zlim = c(-2, 2))
p <- tda_pl(p, "plglob3", lon = "-90,0,90", lat = "-45,0,45", rhs = 1)
plot(p)
}
