% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcmds.R
\name{tda_plotcm}
\alias{tda_plotcm}
\title{Contour plot of a matrix}
\usage{
tda_plotcm(z, levels, ...)
}
\arguments{
\item{z}{value matrix.}

\item{levels}{contour levels.}

\item{...}{passed to \code{\link{tda_run}}.}
}
\value{
path of the PostScript file, invisibly.
}
\description{
TDA's \code{plotcm} over a value matrix at the given levels.
}
\examples{
# a bowl-shaped surface over an 8 x 8 grid: contours at three heights
z <- outer(1:8, 1:8, function(i, j) (i - 4)^2 + (j - 4)^2)
f <- tda_plotcm(z, levels = c(2, 6, 12))
tda_plot_ps(tda_read_ps(f))
}
