% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcmds.R
\name{tda_plotr}
\alias{tda_plotr}
\title{Grey-scale relief of a matrix}
\usage{
tda_plotr(z, grey_range = c(0, 1), ...)
}
\arguments{
\item{z}{value matrix.}

\item{grey_range}{two greys mapped to the value extremes.}

\item{...}{passed to \code{\link{tda_run}}.}
}
\value{
path of the PostScript file, invisibly.
}
\description{
TDA's \code{plotr}.
}
\examples{
# cells shaded from white (smallest value) to black (largest)
f <- tda_plotr(outer(1:5, 1:5, "+"))
tda_plot_ps(tda_read_ps(f))
}
