% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcmds.R
\name{tda_plotsp}
\alias{tda_plotsp}
\title{Akima-smoothed curve through points}
\usage{
tda_plotsp(x, y, ...)
}
\arguments{
\item{x, y}{coordinates.}

\item{...}{passed to \code{\link{tda_run}}.}
}
\value{
path of the PostScript file, invisibly.
}
\description{
TDA's \code{plotsp}.
}
\examples{
# a smooth curve through five points that zigzag between 0 and 1
f <- tda_plotsp(x = c(0, 1, 2, 3, 4), y = c(0, 1, 0, 1, 0))
tda_plot_ps(tda_read_ps(f))
}
