% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/clustering.R, R/descriptive.R,
%   R/graph.R, R/interval.R, R/plotcmds.R, R/rate-methods.R, R/rate.R,
%   R/regression.R, R/smoothing.R, R/spatial.R, R/tda.R
\name{tdaR-methods}
\alias{tdaR-methods}
\alias{print.tda_mds}
\alias{print.tda_freq2}
\alias{print.tda_ragged}
\alias{print.tda_graph}
\alias{print.tda_iv}
\alias{print.tda_ivstat}
\alias{print.tda_ps}
\alias{plot.tda_ps}
\alias{print.summary.tda_fit}
\alias{print.tda_ple}
\alias{summary.tda_ltb}
\alias{print.summary.tda_ltb}
\alias{print.tda_fit}
\alias{print.tda_ltb}
\alias{print.tda_table}
\alias{plot.tda_table}
\alias{print.tda_expr}
\alias{plot.tda_sdvd}
\alias{plot.tda_sd_ps3}
\alias{print.tda_spatial}
\alias{print.tda_result}
\alias{summary.tda_result}
\alias{logLik.tda_result}
\alias{coef.tda_result}
\title{Methods for fitted TDA models}
\usage{
\method{print}{tda_mds}(x, ...)

\method{print}{tda_freq2}(x, ...)

\method{print}{tda_ragged}(x, n = 10L, ...)

\method{print}{tda_graph}(x, ...)

\method{print}{tda_iv}(x, ...)

\method{print}{tda_ivstat}(x, ...)

\method{print}{tda_ps}(x, ...)

\method{plot}{tda_ps}(x, ...)

\method{print}{summary.tda_fit}(x, ...)

\method{print}{tda_ple}(x, n = 10L, ...)

\method{summary}{tda_ltb}(object, n = 10L, ...)

\method{print}{summary.tda_ltb}(x, ...)

\method{print}{tda_fit}(x, ...)

\method{print}{tda_ltb}(x, ...)

\method{print}{tda_table}(x, ...)

\method{plot}{tda_table}(x, ...)

\method{print}{tda_expr}(x, ...)

\method{plot}{tda_sdvd}(x, ...)

\method{plot}{tda_sd_ps3}(x, newpage = TRUE, ...)

\method{print}{tda_spatial}(x, ...)

\method{print}{tda_result}(x, ...)

\method{summary}{tda_result}(object, ...)

\method{logLik}{tda_result}(object, ...)

\method{coef}{tda_result}(object, ...)
}
\arguments{
\item{x, object}{a fitted model or result.}

\item{...}{passed on.}

\item{n}{number of rows to show.}
}
\value{
As for the corresponding generic.
}
\description{
Standard methods for the objects the model functions return. See
\code{\link{tda_rate}} for what those objects are.
}
\examples{
set.seed(41)
d <- data.frame(t = round(rexp(50, 0.1), 1) + 0.5,
                 s = rbinom(50, 1, 0.85), x = rnorm(50))
f <- tda_rate(Surv(t, s) ~ x, d, model = "exponential")
f            # print.tda_fit
coef(f)
summary(f)   # summary.tda_fit / print.summary.tda_fit
}
\keyword{internal}
