% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdar_download_resource.R
\name{tdar_download_resource}
\alias{tdar_download_resource}
\title{Title}
\usage{
tdar_download_resource(id, out_file = NULL, out_dir = ".",
  overwrite = TRUE)
}
\arguments{
\item{out_dir}{}
}
\description{
Title
}
