% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdar_get_ontology.R
\name{tdar_get_ontology}
\alias{tdar_get_ontology}
\title{Title}
\usage{
tdar_get_ontology(ontology)
}
\arguments{
\item{ontology}{Either a path to a tDAR ontology file,
a URL of an ontology resource, or an integer representing
a tDAR ontology.}
}
\description{
Title
}
