% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdar_integration_example_data.R
\docType{data}
\name{tdar_integration_example_data}
\alias{tdar_integration_example_data}
\title{Example tDAR integration output}
\format{A tidy tibble.}
\usage{
tdar_integration_example_data
}
\description{
Example tDAR integration output
}
\keyword{datasets}
